+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - Lucene custom text analysis as an external library 
 - Lucene query parsing as an external library
 - Deploying to the DockerHub
+- `--presearcher` flag to specify [Lucene Monitor Presearcher](https://lucene.apache.org/core/9_1_0/monitor/org/apache/lucene/monitor/Presearcher.html)
 
 ## v2022.05.01